Move PNG utilities to their own module.
