+++ psistega3-core/src/codecs/v1.rs
@@ -557,7 +557,7 @@ impl StegaV1 {
 
         // Write the chunk data to the file. If the chunk
         // is already present then the data will be overwritten.
-        file_utils::insert_or_replace_png_bkgd_chunk(file_path, &chunk)
+        png_utils::insert_or_replace_png_bkgd_chunk(file_path, &chunk)
     }
 
     /// Process the bKGD chunk of a PNG file, and apply any flags that may be present.
@@ -565,7 +565,7 @@ impl StegaV1 {
     /// * `path` - The path to the PNG file.
     ///
     pub fn process_bkgd_chunk(&mut self, path: &str) -> bool {
-        let chunk = if let Some(c) = file_utils::read_png_bkgd_chunk_data(path) {
+        let chunk = if let Some(c) = png_utils::read_png_bkgd_chunk_data(path) {
             c
         } else {
             // This is an error as we should always have a bKGD chunk.
@@ -573,7 +573,7 @@ impl StegaV1 {
         };
 
         // We have a bKGD chunk to process!
-        let data = if let Some(d) = file_utils::get_png_chunk_data(&chunk) {
+        let data = if let Some(d) = png_utils::get_png_chunk_data(&chunk) {
             d
         } else {
             // This is an error as we should always have a bKGD chunk.
@@ -1120,7 +1120,8 @@ mod tests_encode_decode {
         codecs::codec::{Codec, Config},
         hashers,
         utilities::{
-            file_utils::{self, PngChunkType},
+            file_utils,
+            png_utils::{self, PngChunkType},
             test_utils::*,
         },
     };
@@ -1293,7 +1294,7 @@ mod tests_encode_decode {
             .expect("failed to encode the data");
 
         assert!(
-            file_utils::find_png_chunk_start(&output_img_path, PngChunkType::Bkgd).is_some(),
+            png_utils::find_png_chunk_start(&output_img_path, PngChunkType::Bkgd).is_some(),
             "bKGD chunk was not written to the PNG file"
         );
     }
