+++ psistega3-core/src/utilities/mod.rs
@@ -1,5 +1,6 @@
 pub mod file_utils;
 pub mod misc_utils;
+pub mod png_utils;
 
 #[cfg(test)]
 pub mod test_utils;
